// src/api/models/ExamAiEditResponse.ts
export interface ExamAiEditResponse {
    /**
     * 
     * @type {string}
     * @memberof ExamAiEditResponse
     */
    reply: string;
    /**
     * 
     * @type {boolean}
     * @memberof ExamAiEditResponse
     */
    hasEdits: boolean;
    /**
     * 
     * @type {string}
     * @memberof ExamAiEditResponse
     */
    changeSetId: string;
    /**
     * 
     * @type {string}
     * @memberof ExamAiEditResponse
     */
    requestId?: string;
    /**
     * 
     * @type {string}
     * @memberof ExamAiEditResponse
     */
    baseRevision: string;
    /**
     * 
     * @type {string}
     * @memberof ExamAiEditResponse
     */
    resultRevision: string;
    /**
     * 
     * @type {ExamAiEditsEnvelope}
     * @memberof ExamAiEditResponse
     */
    edits: ExamAiEditsEnvelope;
    /**
     * 
     * @type {Array<ExamEditConflict>}
     * @memberof ExamAiEditResponse
     */
    conflicts: Array<ExamEditConflict>;
    /**
     * 
     * @type {string}
     * @memberof ExamAiEditResponse
     */
    checksum: string;
}